refactor: 💡 Splitted the Map Creation in another JS File
LMap.js at root of src
