+++ src/App.js
@@ -1,28 +1,12 @@
-import { useEffect } from 'react';
-import 'ol/ol.css';
-import { Map, View } from 'ol';
-import TileLayer from 'ol/layer/Tile';
-import OSM from 'ol/source/OSM';
-import { fromLonLat } from 'ol/proj';
+import React, { useEffect } from 'react';
+import LMap from './LMap';
 
 function App() {
   useEffect(() => {
     fetch('http://api.open-notify.org/iss-now.json')
       .then((response) => response.json())
       .then((data) => {
-        const issPosition = [Number(data.iss_position.latitude), Number(data.iss_position.longitude)];
-        const map = new Map({
-          target: 'map',
-          layers: [
-            new TileLayer({
-              source: new OSM(),
-            }),
-          ],
-          view: new View({
-            center: fromLonLat([Number(data.iss_position.longitude), Number(data.iss_position.latitude)]),
-            zoom: 5,
-          }),
-        });
+        LMap(data);
       });
   });
